Construct cocoa headless context based on args
